requirements.txt
easyocr>=1.7.2
pyautogui>=0.9.54
Pillow>=11.3.0
numpy>=2.4.4
